pyproject.toml
[project]
name = "campusx-langchain-in-depth"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "arxiv>=2.4.0",
    "chromadb>=1.5.1",
    "docx2txt>=0.9",
    "faiss-cpu>=1.13.2",
    "flask>=3.1.3",
    "grandalf>=0.8",
    "jq>=1.11.0",
    "langchain>=1.2.10",
    "langchain-astradb>=1.0.0",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-experimental>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-openai>=1.1.10",
    "langchain-pinecone>=0.2.13",
    "langchain-tavily>=0.2.17",
    "langgraph>=1.0.9",
    "langgraph-cli[inmem]>=0.4.12",
    "matplotlib>=3.10.8",
    "msoffcrypto-tool>=6.0.0",
    "numpy>=2.4.2",
    "openpyxl>=3.1.5",
    "pandas>=3.0.1",
    "pinecone>=7.3.0",
    "pydantic>=2.12.5",
    "pymupdf>=1.27.1",
    "pypdf>=6.7.1",
    "python-docx>=1.2.0",
    "python-dotenv>=1.2.1",
    "rank-bm25>=0.2.2",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.2.3",
    "standard-imghdr>=3.13.0",
    "streamlit>=1.8.0",
    "tiktoken>=0.12.0",
    "unstructured[docx]>=0.20.8",
    "wikipedia>=1.4.0",
]
